Hilos no están funcionando correctamente

+++ src/interfaces/IRegistro.java
@@ -3,10 +3,10 @@ package interfaces;
 public interface IRegistro {
 	
 	// Reemplaza el valor actual almacenado en el registro con newVal	 
-	public void setValor(byte newVal);
+	public void setValor(int newVal);
 
 	// Devuelve el valor almacenado actualmente en el registro, como un byte
-	public byte getValor();
+	public int getValor();
 	
 	// Borra el contenido del registro.
 	public void clear();
